requirements.txt
pyqt6==6.3.1
python-levenshtein==0.12.2
beautifulsoup4==4.11.1
eyed3==0.9.6
attrs==22.1.0